+++ src/main/java/org/hankki/hankkiserver/api/university/controller/request/UniversitiesPostRequest.java
@@ -1,0 +1,7 @@
+package org.hankki.hankkiserver.api.university.controller.request;
+
+import java.util.List;
+
+public record UniversitiesPostRequest(List<UniversityRequest> universities) {
+  public record UniversityRequest(String name, double longitude, double latitude) { }
+}
