--- src/main/java/org/hankki/hankkiserver/api/university/request/UniversitiesPostRequest.java
@@ -1,7 +1,0 @@
-package org.hankki.hankkiserver.api.university.request;
-
-import java.util.List;
-
-public record UniversitiesPostRequest(List<UniversityRequest> universityRequests) {
-  public record UniversityRequest(String name, double longitude, double latitude) { }
-}
