+++ src/main/java/org/hankki/hankkiserver/api/university/service/command/UniversitiesPostCommand.java
@@ -1,7 +1,7 @@
 package org.hankki.hankkiserver.api.university.service.command;
 
 import java.util.List;
-import org.hankki.hankkiserver.api.university.request.UniversitiesPostRequest;
+import org.hankki.hankkiserver.api.university.controller.request.UniversitiesPostRequest;
 import org.hankki.hankkiserver.domain.common.Point;
 import org.hankki.hankkiserver.domain.university.model.University;
 
@@ -14,7 +14,7 @@ public record UniversitiesPostCommand(List<UniversityCommand> universityCommands
   }
 
   public static UniversitiesPostCommand of(UniversitiesPostRequest universitiesPostRequest) {
-    return new UniversitiesPostCommand(universitiesPostRequest.universityRequests().stream().map(it -> new UniversityCommand(it.name(), it.longitude(), it.latitude())).toList());
+    return new UniversitiesPostCommand(universitiesPostRequest.universities().stream().map(it -> new UniversityCommand(it.name(), it.longitude(), it.latitude())).toList());
   }
 
   public List<University> toEntityList() {
